Rename i18n module to .svelte.ts for proper Svelte rune support
Svelte 5 does not allow exporting $derived from modules, so t() must
remain a function call. Renamed index.ts to index.svelte.ts and updated
all imports to use the .svelte extension.
