+++ src/lib/i18n/index.svelte.ts
